modified directory structure to follow package conventions
